requirements.txt
python-telegram-bot[job-queue]==21.8
python-dotenv==1.0.1
pandas==2.2.3
matplotlib==3.9.2
seaborn==0.13.2
